% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{EXmat}
\alias{EXmat}
\title{Create Covariance Matrix with Exchangeable Correlation Structure}
\usage{
EXmat(time, sigma, rho)
}
\arguments{
\item{time}{vector of time points.}

\item{sigma}{vector of the standard deviations for each time point.}

\item{rho}{correlation coefficient between adjacent time points.}
}
\value{
A \eqn{k} by \eqn{k} covariance matrix with exchangeable correlation structure, with \eqn{k} being the number of the time-points.
}
\description{
Creates a covariance matrix with a common correlation between all time-points. The variances may be heterogeneous.
}
\examples{
EXmat(1:2,rep(1,2),0.5)
K<-EXmat(1:4,c(0.75,0.75,1.5,1.5),0.5)
cov2cor(K)

}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
