% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{F_test}
\alias{F_test}
\title{Multivariate F-test (Hotelling's T2 Test) for Comparing Two Groups.}
\usage{
F_test(est, V, N)
}
\arguments{
\item{est}{vector of estimated mean differences.}

\item{V}{estimate for the covariance matrix.}

\item{N}{number of independent observations.}
}
\value{
A data frame with the test statistic, numerator and denominator degrees of freedom and the p-value.
}
\description{
Tests the multivariate null hypothesis of a zero mean difference between two multivariate (normal) distributions.
}
\details{
The function calculates Hotelling's T2 test for the comparison of two independent mean vectors.
It is assumed that \code{est} is the estimated vector of mean differences and \code{V} is the pooled sample
covariance matrix estimate and \code{N} is the total number of subjects from which these estimates where obtained.
The function is mainly included to allow for validation of the power and sample size
calculations via simulation. See examples.
}
\examples{
library(mvtnorm)
#Assume true difference and some correlation
delta<-c(0.2,0.6)
K<-EXmat(1:2,rep(1,2),0.5)
#Assume 1:2 allocation between the two groups
r<-2/3
#Calculate sample size for 80\% power with significance level 0.05
alpha<-0.05
nn<-n_F_test(power=0.8,r=r,delta=delta,K=K,alpha=alpha)
#Check power
power_F_test(n0=nn$n0,n1=nn$n1,delta=delta,K=K,alpha=alpha)
#Check power by simulation
n0<-ceiling(nn$n0)
n1<-ceiling(nn$n1)
R<-1000
p<-rep(NA,R)
library(mvtnorm)
for(i in 1:R) {
	dat<-sim_1(n0=n0,n1=n1,delta=delta,K=K)
	p[i]<-F_test(dat$est,dat$V,dat$N)$p
}
mean(p<=alpha)

}
\references{
Hotelling, H. (1931). The Generalization of Student's Ratio. Ann. Math. Statist. 2(3): 360-378.
}
\seealso{
\code{\link{n_F_test}}, \code{\link{power_F_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
