% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximum_test.R
\name{max_t_test}
\alias{max_t_test}
\title{Multivariate Maximum t-Test for Comparing Two Groups.}
\usage{
max_t_test(est, V, N, sides = 2)
}
\arguments{
\item{est}{vector of estimated mean differences.}

\item{V}{estimate for the covariance matrix.}

\item{N}{number of independent observations.}

\item{sides}{defines whether the one-sided (sides=1) or the two-sided (sides=2) p-value should be calculated, defaults to sides=2. See details.}
}
\value{
A data frame with the test statistic, numerator and denominator degrees of freedom, the p-value, and an indicator if the p-value is one-sided or two-sided.
}
\description{
Tests the multivariate null hypothesis of a zero mean difference between two multivariate (normal) distributions.
}
\details{
The function calculates the maximum t-test for the comparison of two independent mean vectors.
It is assumed that \code{est} is the estimated vector of mean differences and \code{V} is the pooled sample
covariance matrix estimate and \code{N} is the total number of subjects from which these estimates where obtained.
For the two-sided test,
the test statistic is the absolute value of the maximum of the elementary Wald test statistics. For the one-sided case as currently implemented,
the test statistics is the maximum of the elementary Wald test statistics. Hence the one-sided test is sensitive for deviations from the null hypothesis
corresponding to a positive mean difference. P-values are calculated by assuming a multivariate t-distribution with N-2 degrees of freedom as joint
distribution of the elementary test statistics under the null hypothesis.
The function is mainly included to allow for validation of the power and sample size
calculations via simulation, see examples.
}
\examples{
library(mvtnorm)
#Assume true difference and some correlation
delta<-c(0.2,0.6)
K<-ARmat(1:2,rep(1,2),0.5)
#Assume 1:3 allocation between the two groups
r<-1/4
#Calculate sample size for 80\% power with significance level 0.05
alpha<-0.05
nn<-n_max_t_test(power=0.8,r=r,delta=delta,K=K,alpha=alpha)
#Check power
power_max_t_test(n0=nn$n0,n1=nn$n1,delta=delta,K=K,alpha=alpha)
#Check power by simulation
n0<-nn$n0
n1<-nn$n1
R<-1000
p<-rep(NA,R)
for(i in 1:R) {
 dat<-sim_1(n0=n0,n1=n1,delta=delta,K=K)
 p[i]<-max_t_test(dat$est,dat$V,dat$N)$p
}
mean(p<=alpha)



}
\references{
Hothorn, Torsten, Frank Bretz, and Peter Westfall. "Simultaneous inference in general parametric models." Biometrical Journal: Journal of Mathematical Methods in Biosciences 50.3 (2008): 346-363.
}
\seealso{
\code{\link{n_max_t_test}}, \code{\link{power_max_t_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
