% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{n_F_test}
\alias{n_F_test}
\title{Sample Size Calculation for the Multivariate F-Test (Hotelling's T2 Test)}
\usage{
n_F_test(power = 0.8, r = 0.5, delta, K, alpha = 0.05, interval = c(5, 10000))
}
\arguments{
\item{power}{the aimed for power of the test. The default is power=0.8.}

\item{r}{fraction of subjects in group 1 (\code{n1}). The default value r=0.5 means equally sized groups. For example, a r of 1/3 would mean an allocation of 2:1 for n0:n1.}

\item{delta}{vector of assumed true mean differences.}

\item{K}{assumed true covariance matrix (common to both groups).}

\item{alpha}{significance level, default is 0.05.}

\item{interval}{vector of length two with the interval to search for the sample size, passed to \code{\link[stats]{uniroot}}.}
}
\value{
A data frame with the total sample size \code{N}, the sample sizes in group 0 and group 1, \code{n0} and \code{n1}, the calculated actual power
(which should match closely with the aimed for power), and the indication if a one-sided or two-sided null-hypothesis was assumed.
}
\description{
Calculates the sample size for Hotelling's T2 test for the comparison of two independent mean vectors.
}
\examples{
library(mvtnorm)
#Assume true difference and some correlation
delta<-c(0.2,0.6)
K<-EXmat(1:2,rep(1,2),0.5)
#Assume 1:2 allocation between the two groups
r<-2/3
#Calculate sample size for 80\% power with significance level 0.05
alpha<-0.05
nn<-n_F_test(power=0.8,r=r,delta=delta,K=K,alpha=alpha)
#Check power
power_F_test(n0=nn$n0,n1=nn$n1,delta=delta,K=K,alpha=alpha)
#Check power by simulation
n0<-ceiling(nn$n0)
n1<-ceiling(nn$n1)
R<-1000
p<-rep(NA,R)
library(mvtnorm)
for(i in 1:R) {
	dat<-sim_1(n0=n0,n1=n1,delta=delta,K=K)
	p[i]<-F_test(dat$est,dat$V,dat$N)$p
}
mean(p<=alpha)

}
\references{
Hotelling, Harold. "The Generalization of Student's Ratio" Ann. Math. Statist. 2(3) (1931): 360-378.
}
\seealso{
\code{\link{F_test}}, \code{\link{power_F_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
