% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summen_test.R
\name{n_sum_t_test}
\alias{n_sum_t_test}
\title{Sample size calculation for the Sum test}
\usage{
n_sum_t_test(
  power = 0.8,
  r = 0.5,
  delta,
  K,
  alpha = 0.05,
  sides = 2,
  dftype = c("OBrien", "Logan")[1]
)
}
\arguments{
\item{power}{the aimed for power of the test. The default is power=0.8.}

\item{r}{fraction of subjects in group 1. The default value r=0.5 means equally sized groups.}

\item{delta}{vector of assumed true mean differences.}

\item{K}{assumed true covariance matrix (common to both groups).}

\item{alpha}{significance level, default is 0.05.}

\item{sides}{defines whether the one-sided (sides=1) or the two-sided (sides=2) p-value should be calculated, defaults to sides=2.}

\item{dftype}{type of degrees of freedom. Can be either "OBrien"  (df=N-2k) or "Logan" (df=0.5(N-2)(1-1/k^2)) with k=length(est). Default is dftype="OBrien".}
}
\value{
A data frame with the total sample size \code{N}, the sample sizes in group 0 and group 1, \code{n0} and \code{n1}, the calculated actual power
(which should match closely with the aimed for power), and the indication if a one-sided or two-sided null-hypothesis is tested.
}
\description{
Calculates the sample size for Sum test for the comparison of two independent mean vectors.
}
\examples{
library(mvtnorm)
#Assume true difference and some correlation
delta<-c(0.2,0.6)
K<-ARmat(1:2,rep(1,2),0.5)
#Assume 1:3 allocation between the two groups
r<-1/4
#Calculate sample size for 80\% power with significance level 0.05
alpha<-0.05
nn<-n_sum_t_test(power=0.8,r=r,delta=delta,K=K,alpha=alpha)
#Check power
power_sum_t_test(n0=nn$n0,n1=nn$n1,delta=delta,K=K,alpha=alpha)
#Check power by simulation
n0<-nn$n0
n1<-nn$n1
R<-1000
p<-rep(NA,R)
for(i in 1:R) {
 dat<-sim_1(n0=n0,n1=n1,delta=delta,K=K)
 p[i]<-sum_t_test(dat$est,dat$V,dat$N)$p
}
mean(p<=alpha)

}
\references{
O'Brien, Peter C. "Procedures for comparing samples with multiple endpoints." Biometrics (1984): 1079-1087.
}
\seealso{
\code{\link{sum_t_test}}, \code{\link{power_sum_t_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
