% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_test.R
\name{n_t_test}
\alias{n_t_test}
\title{Sample Size Calculation for the Two-Sample t-test}
\usage{
n_t_test(power = 0.8, r = 0.5, delta, sd_per_group, alpha = 0.05, sides = 2)
}
\arguments{
\item{power}{the aimed for power of the test. The default is power=0.8.}

\item{r}{fraction of subjects in group 1 (\code{n1}). The default value r=0.5 means equally sized groups. For example, a r of 1/3 would mean an allocation of 2:1 for n0:n1.}

\item{delta}{vector of assumed true mean differences.}

\item{sd_per_group}{assumed standard deviation per group. Can be either a number if the same standard deviation is assumed for every group, or a vector with individual standard deviations.}

\item{alpha}{significance level, default is 0.05.}

\item{sides}{defines whether the one-sided (sides=1) or the two-sided (sides=2) p-value should be calculated, defaults to sides=2.}
}
\value{
A data frame with the total sample size \code{N}, the sample sizes in group 0 and group 1, \code{n0} and \code{n1}, the calculated actual power
(which should match closely with the aimed for power), and the indication if a one-sided or two-sided null-hypothesis was assumed.
}
\description{
Calculates the sample size for the Two-Sample t-test for the comparison of two independent mean vectors.
}
\examples{
n_t_test(0.8,0.5,3,c(5,4))
n_t_test(power = 0.8, r = 1/3, delta = 3, sd_per_group = c(5,4))
# linking sample size and power calculations
delta = 3
sd_per_group = 5
nn <- n_t_test(0.8,1/3,delta,sd_per_group)
power_t_test(nn$n0,nn$n1,delta,sd_per_group) # matches the power initiated in n_t_test
}
\seealso{
\code{\link{power_t_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}

Bettina Kroyer \email{bettina.kroyer@meduniwien.ac.at}
}
