% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{power_F_test}
\alias{power_F_test}
\title{Power calculation for the Multivariate F-test (Hotelling's T2 test)}
\usage{
power_F_test(n0, n1 = n0, delta, K, alpha = 0.05)
}
\arguments{
\item{n0}{sample size in group 0.}

\item{n1}{sample size in group 1. Defaults to \code{n0} if not specified otherwise.}

\item{delta}{vector of assumed true mean differences.}

\item{K}{assumed true covariance matrix (common to both groups).}

\item{alpha}{significance level, default is 0.05.}
}
\value{
A data frame with the input sample sizes in group 0 and group 1, \code{n0} and \code{n1}, and the calculated power.
}
\description{
Calculates the power for Hotelling's T2 test for the comparison of two independent mean vectors.
}
\examples{
library(mvtnorm)
#Assume true difference and some correlation
delta<-c(0.2,0.6)
K<-EXmat(1:2,rep(1,2),0.5)
#Assume 1:2 allocation between the two groups
r<-2/3
#Calculate sample size for 80\% power with significance level 0.05
alpha<-0.05
nn<-n_F_test(power=0.8,r=r,delta=delta,K=K,alpha=alpha)
#Check power
power_F_test(n0=nn$n0,n1=nn$n1,delta=delta,K=K,alpha=alpha)
#Check power by simulation
n0<-ceiling(nn$n0)
n1<-ceiling(nn$n1)
R<-1000
p<-rep(NA,R)
library(mvtnorm)
for(i in 1:R) {
	dat<-sim_1(n0=n0,n1=n1,delta=delta,K=K)
	p[i]<-F_test(dat$est,dat$V,dat$N)$p
}
mean(p<=alpha)

}
\references{
Hotelling, Harold. "The generalization of Student's ratio." (1931): 360-378.
}
\seealso{
\code{\link{F_test}}, \code{\link{n_F_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
