% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_test.R
\name{power_t_test}
\alias{power_t_test}
\title{Power Calculation for the Two-sample t-test}
\usage{
power_t_test(n0, n1 = n0, delta, sd_per_group, alpha = 0.05, sides = 2)
}
\arguments{
\item{n0}{sample size in group 0.}

\item{n1}{sample size in group 1. Defaults to \code{n0} if not specified otherwise.}

\item{delta}{assumed true mean difference between groups.}

\item{sd_per_group}{assumed standard deviation per group. Can be either a number if the same standard deviation is assumed for every group, or a vector with individual standard deviations.}

\item{alpha}{significance level, default is 0.05.}

\item{sides}{defines whether the one-sided (sides=1) or the two-sided (sides=2) p-value should be calculated, defaults to sides=2.}
}
\value{
A data frame with the input sample sizes in group 0 and group 1, \code{n0} and \code{n1}, the calculated power, and an indicator if a one-sided or two-sided null-hypothesis was assumed.
.
}
\description{
Calculates the power for the Two-sample t-test for the comparison of two independent mean vectors.
}
\examples{
power_t_test(20,20,3,5,0.05) #sides=2 per default
power_t_test(20,20,3,5,0.05,1)
power_t_test(20,delta=3,sd_per_group=c(2,3))
}
\seealso{
\code{\link{n_t_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}

Bettina Kroyer \email{bettina.kroyer@meduniwien.ac.at}
}
