% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{sim_1}
\alias{sim_1}
\title{Simulate mean and covariances}
\usage{
sim_1(n0 = 100, n1 = 100, delta, K)
}
\arguments{
\item{n0}{sample size in group 0.}

\item{n1}{sample size in group 1.}

\item{delta}{vector of assumed true mean differences.}

\item{K}{assumed true covariance matrix (common to both groups).}
}
\value{
A list with the simulated vector of mean differences (\code{est}),
covariance matrix (\code{V}), the input sample sizes \code{n0} and \code{n1} and the total sample size \code{N=n0+n1}.
}
\description{
Simulates the results of sampling from two independent normal distributions and estimating the mean difference and the pooled covariance matrix.
This is done by drawing mean vectors from multivariate normal distributions and covariance matrices from Wishart distributions and subsequently calculating the difference and
the pooled covariance matrix estimate. The function is intended to support simulations to validate sample size and power calculations performed with
\code{\link{n_F_test}} and \code{\link{power_F_test}}.
}
\references{
Hotelling, Harold. "The generalization of Student's ratio." (1931): 360-378.
}
\seealso{
\code{\link{n_F_test}}, \code{\link{power_F_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
