% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summen_test.R
\name{sum_t_test}
\alias{sum_t_test}
\title{Multivariate Sum Test (O'Brien's Test) for Comparing Two Groups.}
\usage{
sum_t_test(est, V, N, dftype = c("OBrien", "Logan")[1], sides = 2)
}
\arguments{
\item{est}{vector of estimated mean differences.}

\item{V}{estimate for the covariance matrix.}

\item{N}{number of independent observations.}

\item{dftype}{type of degrees of freedom. Can be either "OBrien"  (df=N-2k) or "Logan" (df=0.5(N-2)(1-1/k^2)) with k=length(est). Default is dftype="OBrien".}

\item{sides}{defines whether the one-sided (sides=1) or the two-sided (sides=2) p-value should be calculated, defaults to sides=2. See details.}
}
\value{
A data frame with the test statistic, numerator and denominator degrees of freedom, the p-value, and an indicator if the p-value is one-sided or two-sided.
}
\description{
Tests the multivariate null hypothesis of a zero mean difference between two multivariate (normal) distributions.
}
\details{
Calculates O'Briens OLS test for the comparison of two independent mean vectors.
For the two-sided test, the test statistic of this test is the absolute value of the standardized sum
of elementary Wald test statistics. For the one-sided case as currently implemented,
the test statistics is standardized sum
of elementary Wald test statistics. Hence the one-sided test is sensitive for deviations from the null hypothesis
corresponding to a positive sum of standardized mean differences.
It is assumed that \code{est} is the estimated vector of mean differences and \code{V} is the pooled sample
covariance matrix estimate and \code{N} is the total number of subjects from which these estimates where obtained.
The function is mainly included to allow for validation of the power and sample size
calculations via simulation. See examples.
}
\examples{
library(mvtnorm)
#Assume true difference and some correlation
delta<-c(0.2,0.6)
K<-ARmat(1:2,rep(1,2),0.5)
#Assume 1:3 allocation between the two groups
r<-1/4
#Calculate sample size for 80\% power with significance level 0.05
alpha<-0.05
nn<-n_sum_t_test(power=0.8,r=r,delta=delta,K=K,alpha=alpha)
#Check power
power_sum_t_test(n0=nn$n0,n1=nn$n1,delta=delta,K=K,alpha=alpha)
#Check power by simulation
n0<-nn$n0
n1<-nn$n1
R<-1000
p<-rep(NA,R)
for(i in 1:R) {
 dat<-sim_1(n0=n0,n1=n1,delta=delta,K=K)
 p[i]<-sum_t_test(dat$est,dat$V,dat$N)$p
}
mean(p<=alpha)


}
\references{
O'Brien, Peter C. "Procedures for comparing samples with multiple endpoints." Biometrics (1984): 1079-1087.
Logan, B. R., & Tamhane, A. C. (2004). On O'Brien's OLS and GLS tests for multiple endpoints. Lecture Notes-Monograph Series, 76-88.
}
\seealso{
\code{\link{n_sum_t_test}}, \code{\link{power_sum_t_test}}
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
