% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{ARmat}
\alias{ARmat}
\title{Create covariance matrix of structure AR(1) (autoregressive)}
\usage{
ARmat(time, sigma, rho)
}
\arguments{
\item{time}{vector of time points.}

\item{sigma}{vector of the standard deviations for each time point.}

\item{rho}{correlation coefficient between adjacent time points.}
}
\value{
A covariance matrix of a autoregressive structure with the number of the time points as the dimensions built upon assumed correlation rho between adjacent time points,
standard deviations sigma of every time point.
}
\description{
Create a covariance matrix of structure AR(1) based on given time points, correlation between adjacent time points and standard deviations of each time point.
}
\author{
Robin Ristl \email{robin.ristl@meduniwien.ac.at}
}
